Fix routes and services distinction
