+++ server/services/vacation.service.ts
@@ -1,54 +1,53 @@
-import { Request, Response } from 'express';
-import * as data from './data';
+const data = {
+  vacations: [
+    {
+      id: '10',
+      name: 'Florida beach getaway',
+      description: 'Soft beach sands, warm sunsets, calm waters',
+    },
+    {
+      id: '20',
+      name: 'Glacier excursions',
+      description: 'Vistas of the centuries old glaciers',
+    },
+    {
+      id: '30',
+      name: 'Iceland',
+      description:
+        'Explore the dramatic landscape with volcanoes, geysers, hot springs and lava fields',
+    },
+    {
+      id: '40',
+      name: 'Norway',
+      description: 'Visit breathtaking fjords and waterfalls',
+    },
+    {
+      id: '50',
+      name: 'Rocky mountain vacation',
+      description: 'Long and relaxing hikes through the mountains',
+    },
+  ],
+};
 
-async function getVacations(req: Request, res: Response) {
-  try {
-    const vacations = data.getVacations();
-    res.status(200).json(vacations);
-  } catch (error) {
-    res.status(500).send(error);
-  }
-}
+const addVacation = (vacation) => {
+  vacation.id = `vacation-${vacation.name}`;
+  data.vacations.push(vacation);
+  return vacation;
+};
 
-async function postVacation(req: Request, res: Response) {
-  const vacation = {
-    id: undefined,
-    name: req.body.name,
-    description: req.body.description
-  };
+const updateVacation = (vacation) => {
+  const index = data.vacations.findIndex((v) => v.id === vacation.id);
+  data.vacations.splice(index, 1, vacation);
+  return vacation;
+};
 
-  try {
-    const newVacation = data.addVacation(vacation);
-    res.status(201).json(newVacation);
-  } catch (error) {
-    res.status(500).send(error);
-  }
-}
+const deleteVacation = (id) => {
+  data.vacations = data.vacations.filter((v) => v.id !== id);
+  return true;
+};
 
-async function putVacation(req: Request, res: Response) {
-  const vacation = {
-    id: req.params.id,
-    name: req.body.name,
-    description: req.body.description
-  };
+const getVacations = () => {
+  return data.vacations;
+};
 
-  try {
-    const updatedVacation = data.updateVacation(vacation);
-    res.status(200).json(updatedVacation);
-  } catch (error) {
-    res.status(500).send(error);
-  }
-}
-
-async function deleteVacation(req: Request, res: Response) {
-  const { id } = req.params;
-
-  try {
-    data.deleteVacation(id);
-    res.status(200).json({});
-  } catch (error) {
-    res.status(500).send(error);
-  }
-}
-
-export default { getVacations, postVacation, putVacation, deleteVacation };
+export default { getVacations, addVacation, updateVacation, deleteVacation };
